% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{wkbench_command}
\alias{wkbench_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkbench_command(benchmark, cmd_json)
}
\arguments{
\item{benchmark}{A benchmark handle from \code{\link[=wkbench_new]{wkbench_new()}}.}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
